refactor: illustrator_tools.py を3ファイルに分割（828行→責務分離）
画像生成・プロンプト安全書換・画像リサイズの4責務が混在していた
illustrator_tools.py（828行）を責務ごとに分割:

- prompt_safety.py: プロンプト安全書換（_sanitize_prompt, _rewrite_safe_prompt 等）~260行
- image_processing.py: 画像リサイズ・サムネイル生成（resize_image_variants, _generate_thumbnail）~130行
- illustrator_tools.py: 画像生成・バリデーション（generate_image, validate_image）~340行

循環 import は遅延 import で回避。
テスト86件全通過（ロジック変更なし、ファイル移動 + import 更新のみ）。
